[dev] move UserRepository
