+++ src/main/java/com/carero/repository/UserRepository.java
@@ -1,5 +1,6 @@
-package com.carero.domain.user;
+package com.carero.repository;
 
+import com.carero.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
